# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AERISYN – AI Weather Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /weather/current:
    get:
      operationId: getCurrentWeather
      tags: [weather]
      summary: Get current weather for a location
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
        - name: timezone
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Current weather data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrentWeather"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weather/forecast:
    get:
      operationId: getWeatherForecast
      tags: [weather]
      summary: Get hourly and daily forecast
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
        - name: timezone
          in: query
          required: false
          schema: { type: string }
        - name: days
          in: query
          required: false
          schema: { type: integer, default: 7 }
      responses:
        "200":
          description: Forecast data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForecastData"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weather/historical:
    get:
      operationId: getHistoricalWeather
      tags: [weather]
      summary: Get historical weather data
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
        - name: start_date
          in: query
          required: true
          schema: { type: string, description: "YYYY-MM-DD" }
        - name: end_date
          in: query
          required: true
          schema: { type: string, description: "YYYY-MM-DD" }
        - name: timezone
          in: query
          required: false
          schema: { type: string }
      responses:
        "200":
          description: Historical weather data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HistoricalWeather"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /weather/alerts:
    get:
      operationId: getWeatherAlerts
      tags: [weather]
      summary: Get weather alerts derived from current conditions
      parameters:
        - name: lat
          in: query
          required: true
          schema: { type: number }
        - name: lon
          in: query
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Weather alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WeatherAlert"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]
    CurrentWeather:
      type: object
      required:
        - temperature
        - feels_like
        - humidity
        - pressure
        - wind_speed
        - wind_direction
        - visibility
        - cloud_cover
        - rainfall
        - uv_index
        - weather_code
        - weather_description
        - is_day
        - sunrise
        - sunset
      properties:
        temperature: { type: number }
        feels_like: { type: number }
        humidity: { type: number }
        pressure: { type: number }
        wind_speed: { type: number }
        wind_direction: { type: number }
        wind_gusts: { type: ["number", "null"] }
        visibility: { type: number }
        cloud_cover: { type: number }
        rainfall: { type: number }
        snowfall: { type: ["number", "null"] }
        uv_index: { type: number }
        weather_code: { type: integer }
        weather_description: { type: string }
        is_day: { type: integer }
        sunrise: { type: string }
        sunset: { type: string }
        moonrise: { type: ["string", "null"] }
        moonset: { type: ["string", "null"] }
    HourlyWeather:
      type: object
      required: [time, temperature, humidity, precipitation_probability, weather_code, wind_speed, is_day]
      properties:
        time: { type: string }
        temperature: { type: number }
        feels_like: { type: ["number", "null"] }
        humidity: { type: number }
        precipitation_probability: { type: number }
        precipitation: { type: ["number", "null"] }
        weather_code: { type: integer }
        weather_description: { type: string }
        wind_speed: { type: number }
        pressure: { type: ["number", "null"] }
        uv_index: { type: ["number", "null"] }
        is_day: { type: integer }
    DailyWeather:
      type: object
      required: [date, temp_max, temp_min, weather_code, sunrise, sunset, precipitation_sum, precipitation_probability_max, wind_speed_max]
      properties:
        date: { type: string }
        temp_max: { type: number }
        temp_min: { type: number }
        weather_code: { type: integer }
        weather_description: { type: string }
        sunrise: { type: string }
        sunset: { type: string }
        uv_index_max: { type: ["number", "null"] }
        precipitation_sum: { type: number }
        precipitation_probability_max: { type: number }
        wind_speed_max: { type: number }
        wind_gusts_max: { type: ["number", "null"] }
    ForecastData:
      type: object
      required: [hourly, daily]
      properties:
        hourly:
          type: array
          items:
            $ref: "#/components/schemas/HourlyWeather"
        daily:
          type: array
          items:
            $ref: "#/components/schemas/DailyWeather"
    HistoricalWeather:
      type: object
      required: [daily, hourly]
      properties:
        daily:
          type: array
          items:
            $ref: "#/components/schemas/DailyWeather"
        hourly:
          type: array
          items:
            $ref: "#/components/schemas/HourlyWeather"
    WeatherAlert:
      type: object
      required: [id, type, severity, title, description]
      properties:
        id: { type: string }
        type: { type: string }
        severity:
          type: string
          enum: [low, medium, high, extreme]
        title: { type: string }
        description: { type: string }
        start_time: { type: ["string", "null"] }
        end_time: { type: ["string", "null"] }
...