# packages/api-zod/src/generated/types/forecastData.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * AERISYN – AI Weather Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DailyWeather } from './dailyWeather';
import type { HourlyWeather } from './hourlyWeather';

export interface ForecastData {
  hourly: HourlyWeather[];
  daily: DailyWeather[];
}